requirements.txt
absl-py==1.4.0
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
blis==0.7.9
boto3==1.26.162
botocore==1.29.162
cachetools==5.3.1
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
confection==0.0.4
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
docker-pycreds==0.4.0
docopt==0.6.2
executing==1.2.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.40.0
fsspec==2023.4.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
glob2==0.7
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.56.2
h5py==3.9.0
huggingface-hub==0.15.1
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.0.7
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
keras==2.13.1
kiwisolver==1.4.4
langcodes==3.3.0
libclang==16.0.6
llvmlite==0.40.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.2.1
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==3.0rc1
nltk==3.8.1
numba==0.57.1
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathtools==0.1.2
pathy==0.10.2
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.13
platformdirs==3.8.1
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
Pygments==2.15.1
pynndescent==0.5.10
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thinc==8.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.1.0.dev20230623+cu121
torchaudio==2.1.0.dev20230623+cu121
torchvision==0.16.0.dev20230624+cu121
torchviz==0.0.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typer==0.7.0
typing_extensions==4.4.0
tzdata==2023.3
umap-learn==0.5.3
urllib3==1.26.13
wandb==0.15.5
wasabi==1.1.2
wcwidth==0.2.6
Werkzeug==2.3.6
widgetsnbextension==4.0.8
wrapt==1.15.0
yarg==0.1.9
